FireEye/requirements.txt
# FastAPI + server
fastapi
uvicorn[standard]
python-multipart

# Configuration
pydantic-settings
python-dotenv

# YOLO object detection
ultralytics

# OpenRouter / LLM (OpenAI-compatible SDK)
openai

# Image processing
opencv-python
numpy
Pillow

# Prompt configuration
pyyaml

# Testing
pytest
